pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "benchmark"
version = "0.1.0"
description = "A unified benchmark for mobile manipulation with multiple predictors"
authors = [
    {name = "Kaixin Chai"},
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.23.0,<2.0.0",
    "torch>=2.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "open3d>=0.17.0",
    "imageio>=2.31.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "h5py>=3.6.0",
    "wandb>=0.12.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
]

mobipi = [
    "nerfstudio==1.1.5",
    "scikit-optimize",
    "cma",
    "sentencepiece",
    "peft==0.10.0",
    "transformers==4.36.0",
    "huggingface-hub==0.25.0",
]

lelan = [
    "opencv-python==4.6.0.66",
    "warmup-scheduler",
    "efficientnet-pytorch",
    "openai-clip==1.0.1",
    "scikit-video==1.1.11",
    "lmdb",
    "vit-pytorch",
    "positional-encodings",
]

[project.scripts]
benchmark-run = "scripts.run_benchmark:main"
benchmark-collect-n2m = "scripts.collect_n2m_data:main"
benchmark-collect-mobipi = "scripts.collect_mobipi_images:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["benchmark*"]
exclude = ["env*", "policy*", "predictor*", "data*", "outputs*"]

[tool.setuptools.package-data]
benchmark = ["configs/*.yaml", "configs/**/*.yaml"]
